# perf: Improve module transforms