--- a/packages/babel-helper-module-transforms/package.json
+++ b/packages/babel-helper-module-transforms/package.json
@@ -16,7 +16,6 @@
   "main": "./lib/index.js",
   "dependencies": {
     "@babel/helper-module-imports": "workspace:^",
-    "@babel/helper-simple-access": "workspace:^",
     "@babel/helper-validator-identifier": "workspace:^",
     "@babel/traverse": "workspace:^"
   },