--- a/packages/babel-helper-module-transforms/src/rewrite-live-references.ts
+++ b/packages/babel-helper-module-transforms/src/rewrite-live-references.ts
@@ -1,7 +1,5 @@
-import assert from "assert";
 import { template, types as t } from "@babel/core";
 import type { NodePath, Visitor, Scope } from "@babel/core";
-import simplifyAccess from "@babel/helper-simple-access";
 
 import type { ModuleMetadata } from "./normalize-and-load-metadata.ts";
 
@@ -96,18 +94,6 @@ export default function rewriteLiveReferences(
     rewriteBindingInitVisitor,
     rewriteBindingInitVisitorState,
   );
-
-  // NOTE(logan): The 'Array.from' calls are to make this code with in loose mode.
-  const bindingNames = new Set([
-    ...Array.from(imported.keys()),
-    ...Array.from(exported.keys()),
-  ]);
-  if (process.env.BABEL_8_BREAKING) {
-    simplifyAccess(programPath, bindingNames);
-  } else {
-    // @ts-ignore(Babel 7 vs Babel 8) The third param has been removed in Babel 8.
-    simplifyAccess(programPath, bindingNames, false);
-  }
 
   // Rewrite reads/writes from imports and exports to have the correct behavior.
   const rewriteReferencesVisitorState: RewriteReferencesVisitorState = {
@@ -454,8 +440,6 @@ const rewriteReferencesVisitor: Visitor<RewriteReferencesVisitorState> = {
         const exportedNames = exported.get(localName);
         const importData = imported.get(localName);
         if (exportedNames?.length > 0 || importData) {
-          assert(path.node.operator === "=", "Path was not simplified");
-
           const assignment = path.node;
 
           if (importData) {
@@ -467,14 +451,47 @@ const rewriteReferencesVisitor: Visitor<RewriteReferencesVisitorState> = {
             ]);
           }
 
+          const { operator } = assignment;
+          let newExpr;
+          if (operator === "=") {
+            newExpr = assignment;
+          } else if (
+            operator === "&&=" ||
+            operator === "||=" ||
+            operator === "??="
+          ) {
+            newExpr = t.assignmentExpression(
+              "=",
+              t.cloneNode(assignment.left),
+              t.logicalExpression(
+                operator.slice(0, -1) as t.LogicalExpression["operator"],
+                t.cloneNode(assignment.left) as t.Expression,
+                assignment.right,
+              ),
+            );
+          } else {
+            newExpr = t.assignmentExpression(
+              "=",
+              t.cloneNode(assignment.left),
+              t.binaryExpression(
+                operator.slice(0, -1) as t.BinaryExpression["operator"],
+                t.cloneNode(assignment.left) as t.Expression,
+                assignment.right,
+              ),
+            );
+          }
+
           path.replaceWith(
             buildBindingExportAssignmentExpression(
               this.metadata,
               exportedNames,
-              assignment,
+              newExpr,
               path.scope,
             ),
           );
+
+          path.skip();
+
           requeueInParent(path);
         }
       } else {
@@ -523,9 +540,7 @@ const rewriteReferencesVisitor: Visitor<RewriteReferencesVisitorState> = {
       }
     },
   },
-  "ForOfStatement|ForInStatement"(
-    path: NodePath<t.ForOfStatement | t.ForInStatement>,
-  ) {
+  ForXStatement(path) {
     const { scope, node } = path;
     const { left } = node;
     const { exported, imported, scope: programScope } = this;