--- a/packages/babel-plugin-transform-modules-commonjs/test/fixtures/misc/export-var-and-modify/output.js
+++ b/packages/babel-plugin-transform-modules-commonjs/test/fixtures/misc/export-var-and-modify/output.js
@@ -1,0 +1,22 @@
+"use strict";
+
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+exports.a = void 0;
+var a;
+exports.a = a = a + 1;
+exports.a = a = a - 1;
+exports.a = a = a * 1;
+exports.a = a = a / 1;
+exports.a = a = a % 1;
+exports.a = a = a ** 1;
+exports.a = a = a << 1;
+exports.a = a = a >> 1;
+exports.a = a = a >>> 1;
+exports.a = a = a & 1;
+exports.a = a = a ^ 1;
+exports.a = a = a | 1;
+exports.a = a = a && 1;
+exports.a = a = a || 1;
+exports.a = a = a ?? 1;