… packages/babel-plugin-transform-modules-commonjs/test/fixtures/misc/export-var-and-modify/input.mjs
… yarn.lock