<!--
Before making a PR, please read our contributing guidelines
https://github.com/babel/babel/blob/main/CONTRIBUTING.md

Please note that the Babel Team requires two approvals before merging most PRs.

For issue references: Add a comma-separated list of a [closing word](https://help.github.com/articles/closing-issues-via-commit-messages/) followed by the ticket number fixed by the PR. (it should be underlined in the preview if done correctly)

If you are making a change that should have a docs update: submit another PR to https://github.com/babel/website
-->

| Q                        | A <!--(Can use an emoji 👍) -->
| ------------------------ | ---
| Fixed Issues?            | `Fixes #1, Fixes #2` <!-- remove the (`) quotes and write "Fixes" before the number to link the issues -->
| Patch: Bug Fix?          |
| Major: Breaking Change?  |
| Minor: New Feature?      |
| Tests Added + Pass?      | Yes
| Documentation PR Link    | <!-- If only readme change, add `[skip ci]` to your commits -->
| Any Dependency Changes?  |
| License                  | MIT

<!-- Describe your changes below in as much detail as possible -->
```
all/real-case-mjs-cjs.mjs ts-checker.mjs @ current: 1.08 ops/sec ±1.94% 30 runs (922ms)
all/real-case-mjs-cjs.mjs ts-checker.mjs @ baseline: 0.98 ops/sec ±2.04% 30 runs (1020ms)
all/real-case-mjs-cjs.mjs ts-parser.mjs @ current: 8.68 ops/sec ±5.66% 30 runs (115ms)
all/real-case-mjs-cjs.mjs ts-parser.mjs @ baseline: 8.06 ops/sec ±3.68% 30 runs (124ms)
```
The time was reduced by about 10%